Show previous total_score when calculation fails
Looking at total_score on the dashboard, it frequently flashes to 0 and then back to a calculated value. I'm not sure why this happens but a quick fix is to just store the last good total_score value in total_score_global and show that as a placeholder whenever we calculate that total_score is NaN for some reason. You can see this change working in https://youtu.be/fYZpvWYBMAo

It may be the case that this is happening to other metrics, but I've only noticed it happening for total_score.